YKI(Backend): Move examDate under clerk
