+++ backend/yki/src/main/java/fi/oph/yki/api/clerk/ExamDateController.java
@@ -1,17 +1,20 @@
-package fi.oph.yki.api;
+package fi.oph.yki.api.clerk;
 
 import fi.oph.yki.api.dto.ExamDateDTO;
+import fi.oph.yki.config.ClerkEnabledCondition;
 import fi.oph.yki.service.ExamDateService;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.annotation.Resource;
 import java.util.List;
+import org.springframework.context.annotation.Conditional;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping(value = "/v2/api/exam-date", produces = MediaType.APPLICATION_JSON_VALUE)
+@RequestMapping(value = "/v2/api/clerk/exam-date", produces = MediaType.APPLICATION_JSON_VALUE)
+@Conditional(ClerkEnabledCondition.class)
 public class ExamDateController {
 
   @Resource
